console: Move device data form to components
