+++ pkg/webui/console/components/device-data-form/messages.js
