+++ pkg/webui/console/components/device-data-form/validation-schema.js
