+++ pkg/webui/locales/en.json
@@ -29,6 +29,32 @@
   "console.components.api-key-modal.index.grantedRights": "Granted Rights",
   "console.components.api-key-modal.index.description": "Your API Key has been created successfully.\nNote: After closing this window, the value of the key secret will not be accessible anymore.\nMake sure to copy and store it in a safe place now.",
   "console.components.collaborator-form.index.cannotModifyRightAll": "This user possesses universal admin rights that cannot be modified.",
+  "console.components.device-data-form.messages.activationSettings": "Activation Settings",
+  "console.components.device-data-form.messages.createDevice": "Create Device",
+  "console.components.device-data-form.messages.deviceIdPlaceholder": "my-new-device",
+  "console.components.device-data-form.messages.deviceNamePlaceholder": "My New Device",
+  "console.components.device-data-form.messages.deviceIdDescription": "Unique Identifier of your device; this cannot be changed afterwards",
+  "console.components.device-data-form.messages.deviceNameDescription": "Human friendly name of your device for display purposes",
+  "console.components.device-data-form.messages.deviceDescDescription": "Optional device description; can also be used to save notes about the device",
+  "console.components.device-data-form.messages.joinEUIDescription": "The connected Join EUI",
+  "console.components.device-data-form.messages.leaveBlankPlaceholder": "Leave blank to generate automatically",
+  "console.components.device-data-form.messages.resetsJoinNonces": "Resets Join Nonces",
+  "console.components.device-data-form.messages.resetsFCnt": "Resets Frame Counters",
+  "console.components.device-data-form.messages.deviceEUIDescription": "The device EUI is the unique identifier for this device on the network.",
+  "console.components.device-data-form.messages.nwkKeyDescription": "The encrypted Network Key",
+  "console.components.device-data-form.messages.appKeyDescription": "The App Key is used to secure the communication between your device and the network.",
+  "console.components.device-data-form.messages.appSKeyDescription": "App Session Key",
+  "console.components.device-data-form.messages.fwdNtwkKeyDescription": "Forwarding Network Session Integrity Key (or LoRaWAN 1.0.x NwkSKey)",
+  "console.components.device-data-form.messages.sNtwkSIKeyDescription": "Serving Network Session Integrity Key (only for LoRaWAN 1.1+)",
+  "console.components.device-data-form.messages.ntwkSEncKeyDescription": "Network Session Encryption Key (only for LoRaWAN 1.1+)",
+  "console.components.device-data-form.messages.validate8": "This value needs to be exactly 8 characters long",
+  "console.components.device-data-form.messages.validate16": "This value needs to be exactly 16 characters long",
+  "console.components.device-data-form.messages.validate32": "This value needs to be exactly 32 characters long",
+  "console.components.device-data-form.messages.supportsClassC": "Supports Class C",
+  "console.components.device-data-form.messages.activationMode": "Activation Mode",
+  "console.components.device-data-form.messages.otaa": "Over The Air Activation (OTAA)",
+  "console.components.device-data-form.messages.abp": "Activation By Personalization (ABP)",
+  "console.components.device-data-form.messages.resetWarning": "Reseting is insecure and makes your device susceptible for replay attacks",
   "console.components.gateway-data-form.index.enforced": "Enforced",
   "console.components.gateway-data-form.index.dutyCycle": "Duty Cycle",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "my-new-gateway",
@@ -57,32 +83,6 @@
   "console.containers.api-keys-table.index.keyId": "Key ID",
   "console.containers.api-keys-table.index.grantedRights": "Granted Rights",
   "console.containers.collaborators-table.index.id": "User / Organization ID",
-  "console.containers.device-data-form.messages.activationSettings": "Activation Settings",
-  "console.containers.device-data-form.messages.createDevice": "Create Device",
-  "console.containers.device-data-form.messages.deviceIdPlaceholder": "my-new-device",
-  "console.containers.device-data-form.messages.deviceNamePlaceholder": "My New Device",
-  "console.containers.device-data-form.messages.deviceIdDescription": "Unique Identifier of your device; this cannot be changed afterwards",
-  "console.containers.device-data-form.messages.deviceNameDescription": "Human friendly name of your device for display purposes",
-  "console.containers.device-data-form.messages.deviceDescDescription": "Optional device description; can also be used to save notes about the device",
-  "console.containers.device-data-form.messages.joinEUIDescription": "The connected Join EUI",
-  "console.containers.device-data-form.messages.leaveBlankPlaceholder": "Leave blank to generate automatically",
-  "console.containers.device-data-form.messages.resetsJoinNonces": "Resets Join Nonces",
-  "console.containers.device-data-form.messages.resetsFCnt": "Resets Frame Counters",
-  "console.containers.device-data-form.messages.deviceEUIDescription": "The device EUI is the unique identifier for this device on the network.",
-  "console.containers.device-data-form.messages.nwkKeyDescription": "The encrypted Network Key",
-  "console.containers.device-data-form.messages.appKeyDescription": "The App Key is used to secure the communication between your device and the network.",
-  "console.containers.device-data-form.messages.appSKeyDescription": "App Session Key",
-  "console.containers.device-data-form.messages.fwdNtwkKeyDescription": "Forwarding Network Session Integrity Key (or LoRaWAN 1.0.x NwkSKey)",
-  "console.containers.device-data-form.messages.sNtwkSIKeyDescription": "Serving Network Session Integrity Key (only for LoRaWAN 1.1+)",
-  "console.containers.device-data-form.messages.ntwkSEncKeyDescription": "Network Session Encryption Key (only for LoRaWAN 1.1+)",
-  "console.containers.device-data-form.messages.validate8": "This value needs to be exactly 8 characters long",
-  "console.containers.device-data-form.messages.validate16": "This value needs to be exactly 16 characters long",
-  "console.containers.device-data-form.messages.validate32": "This value needs to be exactly 32 characters long",
-  "console.containers.device-data-form.messages.supportsClassC": "Supports Class C",
-  "console.containers.device-data-form.messages.activationMode": "Activation Mode",
-  "console.containers.device-data-form.messages.otaa": "Over The Air Activation (OTAA)",
-  "console.containers.device-data-form.messages.abp": "Activation By Personalization (ABP)",
-  "console.containers.device-data-form.messages.resetWarning": "Reseting is insecure and makes your device susceptible for replay attacks",
   "console.containers.freq-plans-select.index.freqPlansFetchingFailure": "Could not retrieve the list of available frequency plans",
   "console.containers.webhook-formats-select.index.formatFetchingFailure": "Could not retrieve the list of available webhook formats",
   "console.containers.webhooks-table.index.format": "Format",
