+++ pkg/webui/locales/xx.json
@@ -29,6 +29,32 @@
   "console.components.api-key-modal.index.grantedRights": "Xxxxxxx Xxxxxx",
   "console.components.api-key-modal.index.description": "Xxxx XXX Xxx xxx xxxx xxxxxxx xxxxxxxxxxxx.\nXxxx: Xxxxx xxxxxxx xxxx xxxxxx, xxx xxxxx xx xxx xxx xxxxxx xxxx xxx xx xxxxxxxxxx xxxxxxx.\nXxxx xxxx xx xxxx xxx xxxxx xx xx x xxxx xxxxx xxx.",
   "console.components.collaborator-form.index.cannotModifyRightAll": "Xxxx xxxx xxxxxxxxx xxxxxxxxx xxxxx xxxxxx xxxx xxxxxx xx xxxxxxxx.",
+  "console.components.device-data-form.messages.activationSettings": "Xxxxxxxxxx Xxxxxxxx",
+  "console.components.device-data-form.messages.createDevice": "Xxxxxx Xxxxxx",
+  "console.components.device-data-form.messages.deviceIdPlaceholder": "xx-xxx-xxxxxx",
+  "console.components.device-data-form.messages.deviceNamePlaceholder": "Xx Xxx Xxxxxx",
+  "console.components.device-data-form.messages.deviceIdDescription": "Xxxxxx Xxxxxxxxxx xx xxxx xxxxxx; xxxx xxxxxx xx xxxxxxx xxxxxxxxxx",
+  "console.components.device-data-form.messages.deviceNameDescription": "Xxxxx xxxxxxxx xxxx xx xxxx xxxxxx xxx xxxxxxx xxxxxxxx",
+  "console.components.device-data-form.messages.deviceDescDescription": "Xxxxxxxx xxxxxx xxxxxxxxxxx; xxx xxxx xx xxxx xx xxxx xxxxx xxxxx xxx xxxxxx",
+  "console.components.device-data-form.messages.joinEUIDescription": "Xxx xxxxxxxxx Xxxx XXX",
+  "console.components.device-data-form.messages.leaveBlankPlaceholder": "Xxxxx xxxxx xx xxxxxxxx xxxxxxxxxxxxx",
+  "console.components.device-data-form.messages.resetsJoinNonces": "Xxxxxx Xxxx Xxxxxx",
+  "console.components.device-data-form.messages.resetsFCnt": "Xxxxxx Xxxxx Xxxxxxxx",
+  "console.components.device-data-form.messages.deviceEUIDescription": "Xxx xxxxxx XXX xx xxx xxxxxx xxxxxxxxxx xxx xxxx xxxxxx xx xxx xxxxxxx.",
+  "console.components.device-data-form.messages.nwkKeyDescription": "Xxx xxxxxxxxx Xxxxxxx Xxx",
+  "console.components.device-data-form.messages.appKeyDescription": "Xxx Xxx Xxx xx xxxx xx xxxxxx xxx xxxxxxxxxxxxx xxxxxxx xxxx xxxxxx xxx xxx xxxxxxx.",
+  "console.components.device-data-form.messages.appSKeyDescription": "Xxx Xxxxxxx Xxx",
+  "console.components.device-data-form.messages.fwdNtwkKeyDescription": "Xxxxxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xx XxXxXXX x.x.x XxxXXxx)",
+  "console.components.device-data-form.messages.sNtwkSIKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
+  "console.components.device-data-form.messages.ntwkSEncKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
+  "console.components.device-data-form.messages.validate8": "Xxxx xxxxx xxxxx xx xx xxxxxxx x xxxxxxxxxx xxxx",
+  "console.components.device-data-form.messages.validate16": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
+  "console.components.device-data-form.messages.validate32": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
+  "console.components.device-data-form.messages.supportsClassC": "Xxxxxxxx Xxxxx X",
+  "console.components.device-data-form.messages.activationMode": "Xxxxxxxxxx Xxxx",
+  "console.components.device-data-form.messages.otaa": "Xxxx Xxx Xxx Xxxxxxxxxx (XXXX)",
+  "console.components.device-data-form.messages.abp": "Xxxxxxxxxx Xx Xxxxxxxxxxxxxxx (XXX)",
+  "console.components.device-data-form.messages.resetWarning": "Xxxxxxxx xx xxxxxxxx xxx xxxxx xxxx xxxxxx xxxxxxxxxxx xxx xxxxxx xxxxxxx",
   "console.components.gateway-data-form.index.enforced": "Xxxxxxxx",
   "console.components.gateway-data-form.index.dutyCycle": "Xxxx Xxxxx",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "xx-xxx-xxxxxxx",
@@ -57,32 +83,6 @@
   "console.containers.api-keys-table.index.keyId": "Xxx XX",
   "console.containers.api-keys-table.index.grantedRights": "Xxxxxxx Xxxxxx",
   "console.containers.collaborators-table.index.id": "Xxxx / Xxxxxxxxxxxx XX",
-  "console.containers.device-data-form.messages.activationSettings": "Xxxxxxxxxx Xxxxxxxx",
-  "console.containers.device-data-form.messages.createDevice": "Xxxxxx Xxxxxx",
-  "console.containers.device-data-form.messages.deviceIdPlaceholder": "xx-xxx-xxxxxx",
-  "console.containers.device-data-form.messages.deviceNamePlaceholder": "Xx Xxx Xxxxxx",
-  "console.containers.device-data-form.messages.deviceIdDescription": "Xxxxxx Xxxxxxxxxx xx xxxx xxxxxx; xxxx xxxxxx xx xxxxxxx xxxxxxxxxx",
-  "console.containers.device-data-form.messages.deviceNameDescription": "Xxxxx xxxxxxxx xxxx xx xxxx xxxxxx xxx xxxxxxx xxxxxxxx",
-  "console.containers.device-data-form.messages.deviceDescDescription": "Xxxxxxxx xxxxxx xxxxxxxxxxx; xxx xxxx xx xxxx xx xxxx xxxxx xxxxx xxx xxxxxx",
-  "console.containers.device-data-form.messages.joinEUIDescription": "Xxx xxxxxxxxx Xxxx XXX",
-  "console.containers.device-data-form.messages.leaveBlankPlaceholder": "Xxxxx xxxxx xx xxxxxxxx xxxxxxxxxxxxx",
-  "console.containers.device-data-form.messages.resetsJoinNonces": "Xxxxxx Xxxx Xxxxxx",
-  "console.containers.device-data-form.messages.resetsFCnt": "Xxxxxx Xxxxx Xxxxxxxx",
-  "console.containers.device-data-form.messages.deviceEUIDescription": "Xxx xxxxxx XXX xx xxx xxxxxx xxxxxxxxxx xxx xxxx xxxxxx xx xxx xxxxxxx.",
-  "console.containers.device-data-form.messages.nwkKeyDescription": "Xxx xxxxxxxxx Xxxxxxx Xxx",
-  "console.containers.device-data-form.messages.appKeyDescription": "Xxx Xxx Xxx xx xxxx xx xxxxxx xxx xxxxxxxxxxxxx xxxxxxx xxxx xxxxxx xxx xxx xxxxxxx.",
-  "console.containers.device-data-form.messages.appSKeyDescription": "Xxx Xxxxxxx Xxx",
-  "console.containers.device-data-form.messages.fwdNtwkKeyDescription": "Xxxxxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xx XxXxXXX x.x.x XxxXXxx)",
-  "console.containers.device-data-form.messages.sNtwkSIKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
-  "console.containers.device-data-form.messages.ntwkSEncKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
-  "console.containers.device-data-form.messages.validate8": "Xxxx xxxxx xxxxx xx xx xxxxxxx x xxxxxxxxxx xxxx",
-  "console.containers.device-data-form.messages.validate16": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
-  "console.containers.device-data-form.messages.validate32": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
-  "console.containers.device-data-form.messages.supportsClassC": "Xxxxxxxx Xxxxx X",
-  "console.containers.device-data-form.messages.activationMode": "Xxxxxxxxxx Xxxx",
-  "console.containers.device-data-form.messages.otaa": "Xxxx Xxx Xxx Xxxxxxxxxx (XXXX)",
-  "console.containers.device-data-form.messages.abp": "Xxxxxxxxxx Xx Xxxxxxxxxxxxxxx (XXX)",
-  "console.containers.device-data-form.messages.resetWarning": "Xxxxxxxx xx xxxxxxxx xxx xxxxx xxxx xxxxxx xxxxxxxxxxx xxx xxxxxx xxxxxxx",
   "console.containers.freq-plans-select.index.freqPlansFetchingFailure": "Xxxxx xxx xxxxxxxx xxx xxxx xx xxxxxxxxx xxxxxxxxx xxxxx",
   "console.containers.webhook-formats-select.index.formatFetchingFailure": "Xxxxx xxx xxxxxxxx xxx xxxx xx xxxxxxxxx xxxxxxx xxxxxxx",
   "console.containers.webhooks-table.index.format": "Xxxxxx",
